+++ src/utils/clack-utils.ts
@@ -41,17 +41,29 @@ interface WizardProjectData {
 export interface CliSetupConfig {
   filename: string;
   name: string;
+  gitignore: boolean;
 
   likelyAlreadyHasAuthToken(contents: string): boolean;
   tokenContent(authToken: string): string;
 
   likelyAlreadyHasOrgAndProject(contents: string): boolean;
   orgAndProjContent(org: string, project: string): string;
+
+  likelyAlreadyHasUrl?(contents: string): boolean;
+  urlContent?(url: string): string;
+}
+
+export interface CliSetupConfigContent {
+  authToken: string;
+  org?: string;
+  project?: string;
+  url?: string;
 }
 
 export const rcCliSetupConfig: CliSetupConfig = {
   filename: SENTRY_CLI_RC_FILE,
   name: 'source maps',
+  gitignore: true,
   likelyAlreadyHasAuthToken: function (contents: string): boolean {
     return !!(contents.includes('[auth]') && contents.match(/token=./g));
   },
@@ -70,8 +82,9 @@ export const rcCliSetupConfig: CliSetupConfig = {
   },
 };
 
-export const propertiesCliSetupConfig: CliSetupConfig = {
+export const propertiesCliSetupConfig: Required<CliSetupConfig> = {
   filename: SENTRY_PROPERTIES_FILE,
+  gitignore: true,
   name: 'debug files',
   likelyAlreadyHasAuthToken(contents: string): boolean {
     return !!contents.match(/auth\.token=./g);
@@ -87,6 +100,12 @@ export const propertiesCliSetupConfig: CliSetupConfig = {
   },
   orgAndProjContent(org: string, project: string): string {
     return `defaults.org=${org}\ndefaults.project=${project}`;
+  },
+  likelyAlreadyHasUrl(contents: string): boolean {
+    return !!contents.match(/defaults\.url=./g);
+  },
+  urlContent(url: string): string {
+    return `defaults.url=${url}`;
   },
 };
 
@@ -379,74 +398,141 @@ export async function installPackage({
 }
 
 export async function addSentryCliConfig(
-  authToken: string,
+  { authToken, org, project, url }: CliSetupConfigContent,
   setupConfig: CliSetupConfig = rcCliSetupConfig,
 ): Promise<void> {
   return traceStep('add-sentry-cli-config', async () => {
-    const configExists = fs.existsSync(
-      path.join(process.cwd(), setupConfig.filename),
-    );
-    if (configExists) {
-      const configContents = fs.readFileSync(
-        path.join(process.cwd(), setupConfig.filename),
-        'utf8',
-      );
+    const configPath = path.join(process.cwd(), setupConfig.filename);
+    const configExists = fs.existsSync(configPath);
 
-      if (setupConfig.likelyAlreadyHasAuthToken(configContents)) {
-        clack.log.warn(
-          `${chalk.bold(
-            setupConfig.filename,
-          )} already has auth token. Will not add one.`,
-        );
-      } else {
-        try {
-          await fs.promises.writeFile(
-            path.join(process.cwd(), setupConfig.filename),
-            `${configContents}\n${setupConfig.tokenContent(authToken)}\n`,
-            { encoding: 'utf8', flag: 'w' },
-          );
-          clack.log.success(
-            `Added auth token to ${chalk.bold(
-              setupConfig.filename,
-            )} for you to test uploading ${setupConfig.name} locally.`,
-          );
-        } catch {
-          clack.log.warning(
-            `Failed to add auth token to ${chalk.bold(
-              setupConfig.filename,
-            )}. Uploading ${
-              setupConfig.name
-            } during build will likely not work locally.`,
-          );
-        }
-      }
-    } else {
-      try {
-        await fs.promises.writeFile(
-          path.join(process.cwd(), setupConfig.filename),
-          `${setupConfig.tokenContent(authToken)}\n`,
-          { encoding: 'utf8', flag: 'w' },
-        );
-        clack.log.success(
-          `Created ${chalk.bold(
-            setupConfig.filename,
-          )} with auth token for you to test uploading ${
-            setupConfig.name
-          } locally.`,
-        );
-      } catch {
-        clack.log.warning(
-          `Failed to create ${chalk.bold(
-            setupConfig.filename,
-          )} with auth token. Uploading ${
-            setupConfig.name
-          } during build will likely not work locally.`,
-        );
-      }
+    let configContents =
+      (configExists && fs.readFileSync(configPath, 'utf8')) || '';
+    configContents = addAuthTokenToSentryConfig(
+      configContents,
+      authToken,
+      setupConfig,
+    );
+    configContents = addOrgAndProjectToSentryConfig(
+      configContents,
+      org,
+      project,
+      setupConfig,
+    );
+    configContents = addUrlToSentryConfig(configContents, url, setupConfig);
+
+    try {
+      await fs.promises.writeFile(configPath, configContents, {
+        encoding: 'utf8',
+        flag: 'w',
+      });
+      clack.log.success(
+        `${configExists ? 'Saved' : 'Created'} ${chalk.cyan(
+          setupConfig.filename,
+        )}.`,
+      );
+    } catch {
+      clack.log.warning(
+        `Failed to add auth token to ${chalk.cyan(
+          setupConfig.filename,
+        )}. Uploading ${
+          setupConfig.name
+        } during build will likely not work locally.`,
+      );
     }
 
-    await addAuthTokenFileToGitIgnore(setupConfig.filename);
+    if (setupConfig.gitignore) {
+      await addCliConfigFileToGitIgnore(setupConfig.filename);
+    } else {
+      clack.log.warn(
+        chalk.yellow('DO NOT commit auth token to your repository!'),
+      );
+    }
   });
+}
+
+function addAuthTokenToSentryConfig(
+  configContents: string,
+  authToken: string | undefined,
+  setupConfig: CliSetupConfig,
+): string {
+  if (!authToken) {
+    return configContents;
+  }
+
+  if (setupConfig.likelyAlreadyHasAuthToken(configContents)) {
+    clack.log.warn(
+      `${chalk.cyan(
+        setupConfig.filename,
+      )} already has auth token. Will not add one.`,
+    );
+    return configContents;
+  }
+
+  const newContents = `${configContents}\n${setupConfig.tokenContent(
+    authToken,
+  )}\n`;
+  clack.log.success(
+    `Added auth token to ${chalk.cyan(
+      setupConfig.filename,
+    )} for you to test uploading ${setupConfig.name} locally.`,
+  );
+  return newContents;
+}
+
+function addOrgAndProjectToSentryConfig(
+  configContents: string,
+  org: string | undefined,
+  project: string | undefined,
+  setupConfig: CliSetupConfig,
+): string {
+  if (!org || !project) {
+    return configContents;
+  }
+
+  if (setupConfig.likelyAlreadyHasOrgAndProject(configContents)) {
+    clack.log.warn(
+      `${chalk.cyan(
+        setupConfig.filename,
+      )} already has org and project. Will not add them.`,
+    );
+    return configContents;
+  }
+
+  const newContents = `${configContents}\n${setupConfig.orgAndProjContent(
+    org,
+    project,
+  )}\n`;
+  clack.log.success(
+    `Added default org and project to ${chalk.cyan(
+      setupConfig.filename,
+    )} for you to test uploading ${setupConfig.name} locally.`,
+  );
+  return newContents;
+}
+
+function addUrlToSentryConfig(
+  configContents: string,
+  url: string | undefined,
+  setupConfig: CliSetupConfig,
+): string {
+  if (!url || !setupConfig.urlContent || !setupConfig.likelyAlreadyHasUrl) {
+    return configContents;
+  }
+
+  if (setupConfig.likelyAlreadyHasUrl(configContents)) {
+    clack.log.warn(
+      `${chalk.cyan(setupConfig.filename)} already has url. Will not add one.`,
+    );
+    return configContents;
+  }
+
+  const newContents = `${configContents}\n${setupConfig.urlContent(url)}\n`;
+  clack.log.success(
+    `Added default url to ${chalk.cyan(
+      setupConfig.filename,
+    )} for you to test uploading ${setupConfig.name} locally.`,
+  );
+  return newContents;
 }
 
 export async function addDotEnvSentryBuildPluginFile(
@@ -516,15 +602,15 @@ SENTRY_AUTH_TOKEN="${authToken}"
     }
   }
 
-  await addAuthTokenFileToGitIgnore(SENTRY_DOT_ENV_FILE);
+  await addCliConfigFileToGitIgnore(SENTRY_DOT_ENV_FILE);
 }
 
-async function addAuthTokenFileToGitIgnore(filename: string): Promise<void> {
+async function addCliConfigFileToGitIgnore(filename: string): Promise<void> {
   //TODO: Add a check to see if the file is already ignored in .gitignore
   try {
     await fs.promises.appendFile(
       path.join(process.cwd(), '.gitignore'),
-      `\n# Sentry Auth Token\n${filename}\n`,
+      `\n# Sentry Config File\n${filename}\n`,
       { encoding: 'utf8' },
     );
     clack.log.success(