+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - ref(reactnative): Use clack prompts and share common code (dirty repo, login) (#473)
 - feat(reactnative): Add telemetry (#477)
 - feat(reactnative): Improve `build.gradle` patch so that it's more likely to work without changes in monorepos (#478)
+- fix(reactnative): Save Sentry URL, Organization and Project to `sentry.properties` (#479)
 
 ## 3.15.0